requirements.txt
python-dotenv
langchain
langchain-community
langchain-openai
langchain-chroma
chromadb
tiktoken
pypdf
unstructured
fastapi
python-multipart